fix: 36 compile errors — kernel builds and boots on Linux
- naked_asm!: #[naked] → #[unsafe(naked)], asm! → naked_asm!, drop options(noreturn)
- libm: add dependency + extern crate; replace f32::{sqrt,ln,powi} with libm::{sqrtf,logf,powf}
- KernelEvent::SecurityAlert → AiDecision::SecurityAlert (post_decision not publish)
- VfsError::IoError → VfsError::Io (correct variant name throughout)
- ai_nice_adjust → ai_profile.ai_nice_adjust (field path fix, 3 sites)
- syscall_nr: nr as u32 → nr (already u64, no cast needed)
- path_str.to_owned() → String::from(path_str) (ToOwned not in scope in no_std)
- vmm::unmap_page / pmm::free_frame → memory::unmap_page / memory::free_frame (private mods)
- fingerprint.rs: STATS private → syscall_stats::get_histogram() public accessor
- causal.rs: TASKS private → scheduler::set_woke_by() public accessor
- net.rs: duplicate recv() → recv_vec(); fix key not in scope in tcp_state_machine
- (fd as i32) < 0: add parens to disambiguate cast from generic args
- scripts/run_qemu.sh: image-builder must use --target x86_64-unknown-linux-gnu
- syscall_stats: add get_histogram(pid) public API

+++ kernel/src/ai_engine.rs
@@ -63,7 +63,7 @@ fn apply_decision(decision: AiDecision) {
             // Validate through safety layer, then clamp to live tunable cap.
             let raw   = ai_subsystem::safety::check_scheduler_nice(0, nice_delta);
             let cap   = crate::tunables::AI_NICE_CAP.load(core::sync::atomic::Ordering::Relaxed);
-            let delta = raw.clamp(-cap, cap);
+            let delta = (raw as i32).clamp(-cap, cap) as i8;
             if delta != 0 {
                 crate::scheduler::adjust_priority(pid, delta);
             }
